+++ tests/test_integration_nice_cx_network.py
@@ -23,7 +23,7 @@ SKIP_REASON = 'NDEX2_TEST_SERVER, NDEX2_TEST_USER, NDEX2_TEST_PASS ' \
 
 
 @unittest.skipUnless(os.getenv('NDEX2_TEST_SERVER') is not None, SKIP_REASON)
-class TestClient(unittest.TestCase):
+class TestNiceCXNetworkIntegration(unittest.TestCase):
 
     TEST_DIR = os.path.dirname(__file__)
     WNT_SIGNAL_FILE = os.path.join(TEST_DIR, 'data', 'wntsignaling.cx')